+++ sample/scripts/sample-04.js
@@ -125,6 +125,7 @@ angular.module('sample-04', ['adf', 'LocalStorageModule'])
   $scope.name = name;
   $scope.model = model;
   $scope.collapsible = false;
+  $scope.maximizable = false;
 
   $scope.$on('adfDashboardChanged', function(event, name, model) {
     localStorageService.set(name, model);